Made Changes to conveyorbelts
Merged multiple scripts into one. Changed removed other conveyor belts.
Changed ROtation variables on new conveyorbelt, and made multiple variants.
