+++ MicrowaveGame2D/Assets/Prefabs/Hazards/Conveyor Belt(Down).prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 1790528090271822784}
   - component: {fileID: 1790528090271822787}
   - component: {fileID: 1790528090271822786}
-  - component: {fileID: 1790528090271822796}
+  - component: {fileID: 741581926591845417}
   m_Layer: 0
   m_Name: Conveyor Belt(Down)
   m_TagString: Untagged
@@ -26,13 +26,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1790528090271822799}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -2, y: 3, z: 0}
+  m_LocalRotation: {x: 0, y: 0, z: 1, w: 0}
+  m_LocalPosition: {x: -2, y: -1.1, z: 0}
   m_LocalScale: {x: 1, y: -4, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 180}
 --- !u!212 &1790528090271822787
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -110,7 +110,7 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.971956, y: 0.9860898}
   m_EdgeRadius: 0
---- !u!114 &1790528090271822796
+--- !u!114 &741581926591845417
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
