+++ MicrowaveGame2D/Assets/Prefabs/Hazards/Conveyor Belt(Left) Variant.prefab
@@ -1,0 +1,59 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &5011356794199222241
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 4.26
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -1.39
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: -0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822784, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 270
+      objectReference: {fileID: 0}
+    - target: {fileID: 1790528090271822799, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
+      propertyPath: m_Name
+      value: Conveyor Belt(Left)
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: bf247bb12ef62a74ba6807eae80c6343, type: 3}
