+++ MicrowaveGame2D/Assets/Prefabs/Hazards/Conveyor Belt(Left) Variant.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 719223a2a0a0d834591946cff4cc7d83
+guid: 8c3f042bee9a5284e94b3dc69df3e97a
 PrefabImporter:
   externalObjects: {}
   userData:
