+++ MicrowaveGame2D/Assets/Prefabs/Hazards/Conveyor Belt(Right) Variant.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6aade1b9c3aebab44a2757ac82afdaa9
+guid: 0909d049cf2f55b40ae5e9f754e646c7
 PrefabImporter:
   externalObjects: {}
   userData:
