+++ MicrowaveGame2D/Assets/Prefabs/Hazards/Conveyor Belt(Up) Variant.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fe09ecc6938b44c4e88b5d616c14b2a8
+guid: dd391ef606ae8424d916b3fab283bd92
 PrefabImporter:
   externalObjects: {}
   userData:
