--- MicrowaveGame2D/Assets/Scripts/Hazards/ConveyorBeltLeft.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 77be19ad8f7a9f34281240e8de205ee9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
