--- MicrowaveGame2D/Assets/Scripts/Hazards/ConveyorBeltRight.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 271116ebd5730d349a102273c868b9a2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
