--- MicrowaveGame2D/Assets/Scripts/Hazards/ConveyorBeltUp.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bf9ffd2df314f74586a51a4ca763460
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
